requirements.txt
faiss-cpu==1.9.0.post1
Flask==3.1.0
numpy==2.2.2
python-dotenv==1.0.1
sentence-transformers==3.3.1
